@@ -15,16 +15,16 @@
 		"format": "prettier --write ."
 	},
 	"devDependencies": {
-		"@sveltejs/adapter-auto": "^3.0.0",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
+		"@sveltejs/adapter-auto": "^7.0.1",
+		"@sveltejs/kit": "^2.5.27",
+		"@sveltejs/vite-plugin-svelte": "^6.2.4",
 		"prettier": "^3.1.1",
 		"prettier-plugin-svelte": "^3.4.1",
-		"svelte": "^4.2.7",
+		"svelte": "^5.51.2",
 		"svelte-check": "^4.0.0",
-		"typescript": "^5.0.0",
-		"vite": "^5.0.3",
-		"vitest": "^2.0.0"
+		"typescript": "^5.5.0",
+		"vite": "^7.3.1",
+		"vitest": "^4.0.18"
 	},
 	"type": "module",
 	"dependencies": {
